requirements.txt
langchain==0.1.13
langchain-community==0.0.36
langchain-openai==0.1.1
python-dotenv==1.0.0
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
chromadb==0.4.21
sentence-transformers==2.2.2
openai>=1.10.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4>=4.12.0
ipython==8.15.0
jupyter==1.0.0
streamlit==1.28.0
streamlit-chat==0.1.1
networkx>=3.0
youtube-transcript-api>=0.6.0
